% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_data.R
\name{raster_multiple_bands_plots}
\alias{raster_multiple_bands_plots}
\title{Visualization functions for raster data, multiple bands}
\usage{
raster_multiple_bands_plots(
  raster_data,
  RGB_bands,
  indices_formula,
  band_order,
  date_order,
  n_bands_explore,
  n_first_date,
  x_order,
  y_order,
  x_value,
  y_value
)
}
\arguments{
\item{x_value, }{ex: NULL, or integer: 653570}

\item{y_value, }{ex: NULL, or integer: 5192770}

\item{raster_data:}{standardized raster data as stars objects}

\item{RGB_bands:}{list of bands corresponding to Red, Green and Blue bands, ex: c("B4", "B3", "B2")}

\item{indices_formula:}{list of index name and its string formula, ex: c("NDSI" = "(B3 - B11) / (B3 + B11)", "NDVI" = "(B8 - B4) / (B8 + B4)")}

\item{band_order:}{order of band in band_names, ex: 1}

\item{date_order:}{order of date in time_dates, ex: 1}

\item{n_bands_explore:}{number of bands to explore, ex: 6}

\item{n_first_date:}{number of dates to explore, ex: 6}

\item{x_order:}{order of coordinate x, ex: 1}

\item{y_order:}{order of coordinate y, ex: 1}
}
\value{
plot functions
}
\description{
Visualization functions for raster data, multiple bands
}
\examples{
plot_func <- raster_multiple_bands_plots(raster_data, x_order = 100, y_order = 50)
plot_func <- raster_multiple_bands_plots(raster_data,
                                         RGB_bands = c("B4", "B3", "B2"),
                                         indices_formula = c("NDSI" = "(B3 - B11) / (B3 + B11)",
                                                             "NDVI" = "(B8 - B4) / (B8 + B4)"),
                                         band_order = 1, date_order = 1,
                                         n_bands_explore = 6, n_first_date = 6,
                                         x_order = 1, y_order = 1, x_value = NULL, y_value = NULL,)
}
