% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{raster_multiple_bands_report_combine}
\alias{raster_multiple_bands_report_combine}
\title{Create a pdf report for raster multiple bands data}
\usage{
raster_multiple_bands_report_combine(
  data_path,
  suffix = ".tif",
  band_names = c("B2", "B3", "B4", "B8", "B11", "B12"),
  RGB_bands = c("B4", "B3", "B2"),
  indices_formula = c(NDSI = "(B3 - B11) / (B3 + B11)", NDVI = "(B8 - B4) / (B8 + B4)"),
  band_order = 1,
  date_order = 1,
  n_bands_explore = 6,
  n_first_date = 6,
  x_order = 1,
  y_order = 1,
  x_value = NULL,
  y_value = NULL,
  output_path = "output/raster_multiple_bands_report.pdf"
)
}
\arguments{
\item{data_path, }{ex: 'data/sentinel2'}

\item{x_value, }{ex: NULL, or integer: 653570}

\item{y_value, }{ex: NULL, or integer: 5192770}

\item{suffix:}{file type, ex: '.tif'}

\item{band_names:}{list of bands, ex: c("B2", "B3", "B4", "B8", "B11", "B12")}

\item{RGB_bands:}{list of bands corresponding to Red, Green and Blue bands, ex: c("B4", "B3", "B2")}

\item{indices_formula:}{list of index name and its string formula, ex: c("NDSI" = "(B3 - B11) / (B3 + B11)", "NDVI" = "(B8 - B4) / (B8 + B4)")}

\item{band_order:}{order of band in band_names, ex: 1}

\item{date_order:}{order of date in time_dates, ex: 1}

\item{n_bands_explore:}{number of bands to explore, ex: 6}

\item{n_first_date:}{number of dates to explore, ex: 6}

\item{x_order:}{order of coordinate x, ex: 1}

\item{y_order:}{order of coordinate y, ex: 1}

\item{output_path:}{output dir and name of report, ex: 'output/raster_multiple_bands_report.pdf'}
}
\value{
a pdf report
}
\description{
Create a pdf report for raster multiple bands data
}
\examples{
raster_multiple_bands_report_combine(data_path, suffix = '.tif', band_names = c("B2", "B3", "B4", "B8", "B11", "B12"),
                                     RGB_bands = c("B4", "B3", "B2"),
                                     indices_formula = c("NDSI" = "(B3 - B11) / (B3 + B11)",
                                                         "NDVI" = "(B8 - B4) / (B8 + B4)"),
                                     band_order = 1, date_order = 1,
                                     n_bands_explore = 6, n_first_date = 6,
                                     x_order = 100, y_order = 50, x_value = NULL, y_value = NULL,
                                     output_path='output/raster_multiple_bands_report.pdf')
}
