% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_data.R
\name{tabular_single_value_plots}
\alias{tabular_single_value_plots}
\title{Visualization functions for tabular data, single value}
\usage{
tabular_single_value_plots(data, cname, time_frequency = "daily")
}
\arguments{
\item{data:}{standardized data}

\item{cname:}{short name representing for value, ex: 'co2', 'temp'}

\item{time_frequency:}{time frequency, ex: 'daily', 'monthly', 'yearly'}
}
\value{
plot functions
}
\description{
Visualization functions for tabular data, single value
}
\examples{
plot_func <- tabular_single_value_plots(data, 'temp', time_frequency = 'daily')
p <- plot_func$line_plot()
p
}
